pyproject.toml
[project]
name = "baulog"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.136.1",
    "langchain>=1.2.15",
    "langchain-google-genai>=1.0.0",
    "google-generativeai>=0.8.0",
    "python-dotenv>=1.0.0",
    "email-validator>=2.0.0",
]
